+++ ChangeLog.md
@@ -9,6 +9,9 @@ More detailed information about incremental changes can be found in the
 
 ### 3.14.7 (2022-xx-yy)
 
+- Fixed that ComputeSensitivityMatrix() of sIpopt assumed that there are
+  no more than two parameters [#578, by Andrea Vescovini].
+
 ### 3.14.6 (2022-05-02)
 
 - Fixed mapping of meta data for inequalities, e.g., constraint names,